% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/curleytheme.R
\name{curleytheme}
\alias{curleytheme}
\title{Curley Lab ggplot2 generic theme}
\usage{
curleytheme()
}
\value{
Generic theme for ggplots
}
\description{
Curley Lab ggplot2 generic theme
}
\section{Further details}{

A generic plotting theme for ggplot graphs
}
\examples{
ggplot(df, aes(x, y)) + geom_point() + curleytheme
}

