% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expandrows.R
\name{expandrows}
\alias{expandrows}
\title{Expand rows of google forms}
\usage{
expandrows(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
The dataframe with expanded rows, as well as a score variable.
}
\description{
Expand rows of google forms
}
\section{Further details}{

Expand rows that have a variable where 2 or more factors
are split by commas into separate rows with
all other variables duplicated.
Make sure columns/variables are named "Actor" and "Recipient".
The score variable will be '1' for Actors that are clear winners
and a '0.5' for ties between Actors and Recipients.
}
\examples{
expandrows(df)
}

